% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testeFelipe.R
\encoding{UTF-8}
\name{testeFelipe}
\alias{testeFelipe}
\title{Tamanho da Amostra - 2 Fatores n observações}
\usage{
testeFelipe(alpha, a, b, ate, betas, sigma)
}
\arguments{
\item{alpha}{Erro Tipo I esperado}

\item{a}{Tamanho do Fator 1}

\item{b}{Tamanho do Fator 2}

\item{ate}{Cálcular o erro Tipo II de n=2 até o valor especificado neste parâmetro}

\item{betas}{Vetor contendo o valor dos Betas (ou distância)}

\item{sigma}{Desvio-padrão}
}
\value{
Matrix contendo  \code{n}, \code{a}, \code{b}, \code{df1}, \code{df2}, \code{f_crit}, \code{phi_2}, \code{errotip2} e \code{poder_do_teste}.
}
\description{
Função que calcula o Erro Tipo II e o Poder do Teste em um experimento com 2 fatores e n observações.
}
\details{
Parâmetro de não-centralidade:
\deqn{\phi^2=\frac{an\sum_{j=1}^b\beta_i^2}{\sigma^2}}
Erro tipo II:
\deqn{\beta(\phi^2)=P(F\leq F_{crítico}|F_{(df1=b-1,\;df2=ab(n-1),\;npc=\phi^2)})} 
Poder do teste:
\deqn{1-\beta(\phi^2)}
}
\examples{

##Page 202 - Design and Analysis of Experiments 8th by DOUGLAS C. MONTGOMERY: 
testeFelipe(alpha=0.05,a=50,b=5,ate=50,betas=c(-20,20),sigma=25)

}
\references{
MONTGOMERY, Douglas C. Design and Analysis of Experiments. 8. ed
}
\seealso{
\code{\link[stats]{qf}}, \code{\link[stats]{pf}}
}
\author{
Felipe da Rocha Ferreira
}
